+++ src/main/java/com/mflechl/mfchess/Notation.java
@@ -21,13 +21,13 @@ class Notation extends JLabel {
         notationStrings.add(""); //entry for currentStaticState of new board
     }
 
-    static String getMoveNotation(IBoard board, State state, State prevState, int fromLine, int fromRow, int toLine, int toRow, int movingPiece, int eliminatedPiece, SpecialMove sMove) {
+    static String getMoveNotation(IBoard board, State state, State prevState, byte fromLine, byte fromRow, byte toLine, byte toRow, byte movingPiece, byte eliminatedPiece, SpecialMove sMove) {
         return getMoveNotation(board, state.moveNumber, state.check, fromLine, fromRow, toLine, toRow, movingPiece, eliminatedPiece, sMove.enPassant, sMove.castling, prevState, state.mate, state.remis);
     }
 
-    static String getMoveNotation(IBoard board, int imove, boolean check, int aLine, int aRow, int toLine, int toRow, int piece, int eliminatedPiece, boolean enpassant, boolean castling, State state, boolean mate, boolean remis) {
+    static String getMoveNotation(IBoard board, int imove, boolean check, byte aLine, byte aRow, byte toLine, byte toRow, byte piece, byte eliminatedPiece, boolean enpassant, boolean castling, State state, boolean mate, boolean remis) {
         String lbl = "";
-        int apiece = Math.abs(piece);
+        byte apiece = (byte) Math.abs(piece);
 
         //write move number
         if (piece > 0) lbl += imove + ". ";
@@ -76,11 +76,11 @@ class Notation extends JLabel {
         return lbl;
     }
     //lines=normal notation minus 1; rows=a->1, b->2, ...; ep=en passant done
-    void addMove(IBoard board, State state, State prevState, int fromLine, int fromRow, int toLine, int toRow, int movingPiece, int eliminatedPiece, SpecialMove sMove) {
+    void addMove(IBoard board, State state, State prevState, byte fromLine, byte fromRow, byte toLine, byte toRow, byte movingPiece, byte eliminatedPiece, SpecialMove sMove) {
         addMove(board, state.nMoves, state.moveNumber, state.check, fromLine, fromRow, toLine, toRow, movingPiece, eliminatedPiece, sMove.enPassant, sMove.castling, prevState, state.mate, state.remis);
     }
 
-    void addMove(IBoard board, int pos, int imove, boolean check, int fromLine, int fromRow, int toLine, int toRow, int piece, int eliminatedPiece, boolean enpassant, boolean castling, State state, boolean mate, boolean remis) {
+    void addMove(IBoard board, int pos, int imove, boolean check, byte fromLine, byte fromRow, byte toLine, byte toRow, byte piece, byte eliminatedPiece, boolean enpassant, boolean castling, State state, boolean mate, boolean remis) {
         String lbl = getMoveNotation(board, imove, check, fromLine, fromRow, toLine, toRow, piece, eliminatedPiece, enpassant, castling, state, mate, remis);
         if (pos >= 0) updateText(lbl, pos);
     }
@@ -116,11 +116,11 @@ class Notation extends JLabel {
 
     //check if the notation is ambiguous, i.e. if row and/or line need also be given
     //board is the board *before* the last move has happened
-    static String ambiguity(IBoard board, int piece, int fromLine, int fromRow, int toLine, int toRow, State state) {
+    static String ambiguity(IBoard board, byte piece, byte fromLine, byte fromRow, byte toLine, byte toRow, State state) {
         String amb = "";
-        int namb = 0;
-        for (int iLine = 0; iLine < 8; iLine++) {
-            for (int iRow = 0; iRow < 8; iRow++) {
+        byte namb = 0;
+        for (byte iLine = 0; iLine < 8; iLine++) {
+            for (byte iRow = 0; iRow < 8; iRow++) {
                 if (iLine == fromLine && iRow == fromRow) continue; //this is the move that was actually made
                 if (board.setup[iLine][iRow] == piece) {
                     //System.out.println("ambiguity: state="+state+"sDummy: "+Move.sDummy+"\n"+board);
