+++ src/main/java/com/mflechl/mfchess/NotationToState.java
@@ -12,26 +12,26 @@ public final class NotationToState {
     private static final String patternMoveNumber = "[0-9]*\\.";
     private static final String patternPiece = "^[KQRBN].*";
 
-    private static final Map<String, Integer> pieceMap = new HashMap<>();
+    private static final Map<String, Byte> pieceMap = new HashMap<>();
     static {
-        pieceMap.put("K", 1);
-        pieceMap.put("Q", 2);
-        pieceMap.put("R", 3);
-        pieceMap.put("B", 4);
-        pieceMap.put("N", 5);
+        pieceMap.put("K", (byte) 1);
+        pieceMap.put("Q", (byte) 2);
+        pieceMap.put("R", (byte) 3);
+        pieceMap.put("B", (byte) 4);
+        pieceMap.put("N", (byte) 5);
     }
 
-    private static final Map<String, Integer> rowMap = new HashMap<>();
+    private static final Map<String, Byte> rowMap = new HashMap<>();
 
     static {
-        rowMap.put("a", 0);
-        rowMap.put("b", 1);
-        rowMap.put("c", 2);
-        rowMap.put("d", 3);
-        rowMap.put("e", 4);
-        rowMap.put("f", 5);
-        rowMap.put("g", 6);
-        rowMap.put("h", 7);
+        rowMap.put("a", (byte) 0);
+        rowMap.put("b", (byte) 1);
+        rowMap.put("c", (byte) 2);
+        rowMap.put("d", (byte) 3);
+        rowMap.put("e", (byte) 4);
+        rowMap.put("f", (byte) 5);
+        rowMap.put("g", (byte) 6);
+        rowMap.put("h", (byte) 7);
     }
 
     static IBoardState noteToBoard(final String move, IBoardState prevBoard) {
@@ -42,11 +42,11 @@ public final class NotationToState {
         String str = move.replaceAll("^\\s+", "").replaceAll("\\s$", "");
         IBoardState currBoard = new IBoardState(prevBoard);
 
-        int toLine, toRow;
-        int fromLine = -1, fromRow = -1;
-        int piece;
-        int toPiece = -1;
-        int turnOf = prevBoard.state.turnOf;
+        byte toLine, toRow;
+        byte fromLine = -1, fromRow = -1;
+        byte piece;
+        byte toPiece = -1;
+        byte turnOf = prevBoard.state.turnOf;
         SpecialMove sDummy;
         SpecialMove sMove = new SpecialMove();
 
@@ -56,7 +56,7 @@ public final class NotationToState {
             if (turnOf == ChessBoard.WHITE) fromLine = 0;
             else fromLine = 7;
             toLine = fromLine;
-            piece = ChessBoard.KING * turnOf;
+            piece = (byte) (ChessBoard.KING * turnOf);
 
             if (str.startsWith("0-0-0")) { //queen-side castling
                 toRow = rowMap.get("c");
@@ -68,38 +68,39 @@ public final class NotationToState {
                 return null;
         } else {
             //which piece moved?
-            piece = turnOf * ChessBoard.PAWN;
+            piece = (byte) (turnOf * ChessBoard.PAWN);
             if (str.matches(patternPiece)) {
-                piece = turnOf * pieceMap.get(str.substring(0, 1));
+                piece = (byte) (turnOf * pieceMap.get(str.substring(0, 1)));
                 str = str.substring(1);                                       //remove piece from string
             }
 
-            int remFromRow = -1, remFromLine = -1;
+            byte remFromRow = -1, remFromLine = -1;
             if (str.matches("[abcdefgh]x?[abcdefgh].*")) {
                 remFromRow = rowMap.get(str.substring(0, 1));
                 str = str.substring(1);                                       //remove piece from string
             } else if (str.matches("\\dx?[abcdefgh].*")) {
-                remFromLine = Integer.parseInt(str.substring(0, 1)) - 1;
+                remFromLine = (byte) (Byte.parseByte(str.substring(0, 1)) - 1);
                 str = str.substring(1);                                       //remove piece from string
             } else if (str.matches("[abcdefgh]\\dx?[abcdefgh].*")) {
                 remFromRow = rowMap.get(str.substring(0, 1));
-                remFromLine = Integer.parseInt(str.substring(1, 2)) - 1;
+                remFromLine = (byte) (Byte.parseByte(str.substring(1, 2)) - 1);
                 str = str.substring(2);                                       //remove piece from string
             }
             if (str.substring(0, 1).equals("x")) str = str.substring(1);     //remove "x"; str is now final state
 
             //get TO coordinates and remove info from string
-            toLine = Integer.parseInt(str.substring(1, 2)) - 1;
+            toLine = (byte) (Byte.parseByte(str.substring(1, 2)) - 1);
             toRow = rowMap.get(str.substring(0, 1));
 
             //promotion: to which piece?
             if (str.length() > 2) {
-                if (str.substring(2, 3).matches(patternPiece)) toPiece = turnOf * pieceMap.get(str.substring(2, 3));
+                if (str.substring(2, 3).matches(patternPiece))
+                    toPiece = (byte) (turnOf * pieceMap.get(str.substring(2, 3)));
             }
 
             //get FROM coordinates
-            for (int iLine = 0; iLine < 8; iLine++) {
-                for (int iRow = 0; iRow < 8; iRow++) {
+            for (byte iLine = 0; iLine < 8; iLine++) {
+                for (byte iRow = 0; iRow < 8; iRow++) {
                     if (piece != currBoard.setup[iLine][iRow]) continue;
                     if (remFromRow >= 0 && iRow != remFromRow) continue;
                     if (remFromLine >= 0 && iLine != remFromLine) continue;
